round of corrections after kyles/bens/coreys comments
