+++ src/sparsezoo/model/__init__.py
@@ -14,27 +14,8 @@
 
 # flake8: noqa
 
-import os
+from .model import *
+from .utils import *
 
 
-BASE_API_URL = (
-    os.getenv("SPARSEZOO_API_URL")
-    if os.getenv("SPARSEZOO_API_URL")
-    else "https://api.neuralmagic.com"
-)
-MODELS_API_URL = f"{BASE_API_URL}/models"
-
-from .authentication import *
-from .download import *
-from .helpers import *
-from .numpy import *
-from .requests import *
-
-
-__all__ = (
-    helpers.__all__
-    + authentication.__all__
-    + download.__all__
-    + numpy.__all__
-    + requests.__all__
-)
+__all__ = utils.__all__ + model.__all__
