+++ src/sparsezoo/objects/__init__.py
@@ -12,19 +12,11 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-# flake8: noqa
-
+from .directories import *
 from .directory import *
+
+# flake8: noqa
 from .file import *
-from .model_helpers import *
-from .model_objects import *
-from .utils import *
 
 
-__all__ = (
-    file.__all__
-    + directory.__all__
-    + model_objects.__all__
-    + utils.__all__
-    + model_helpers.__all__
-)
+__all__ = file.__all__ + directory.__all__ + directories.__all__
