+++ src/sparsezoo/utils/download.py
@@ -19,7 +19,7 @@ from typing import Iterator, NamedTuple, Union
 import requests
 from tqdm import auto, tqdm, tqdm_notebook
 
-from .utils import clean_path, create_parent_dirs
+from .helpers import clean_path, create_parent_dirs
 
 
 __all__ = ["download_file"]
