+++ src/sparsezoo/validation/validator.py
@@ -16,14 +16,13 @@
 Validator class that checks, whether the contents of ModelDirectory
 class objects are valid
 """
-from __future__ import annotations
+from __future__ import annotations  # noqa F407
 
 import os
 from typing import TYPE_CHECKING, Callable, Dict, Optional, Set, Tuple, Union
 
 from sparsezoo.objects import Directory, File
-
-from .integrations import (
+from sparsezoo.validation import (
     validate_cv_classification,
     validate_cv_detection,
     validate_cv_segmentation,
