+++ tests/sparsezoo/model/test_model.py
