+++ tests/sparsezoo/objects/test_directories.py
