+++ tests/sparsezoo/objects/test_directory.py
@@ -20,8 +20,7 @@ import tempfile
 import pytest
 
 from sparsezoo.objects import Directory, File
-
-from .test_file import _create_sample_file
+from tests.sparsezoo.objects.test_file import _create_sample_file
 
 
 def _create_files_directory(files_extensions, temp_dir):
