+++ tests/sparsezoo/search/test_search.py
